backend/requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-magic>=0.4.27
nltk>=3.8.1
spacy>=3.7.0
pdf2image>=1.16.3
PyPDF2>=3.0.1
python-docx>=1.0.0
pytesseract>=0.3.10
scikit-learn>=1.3.0
